Split out Routing.ts into smaller files
